requirements.txt
yt-dlp
moviepy
openai-whisper
spacy
sentence-transformers
faiss-cpu
opencv-python
groq
gradio==4.37.2